+++ tests/data/__init__.py
@@ -1,2 +1,3 @@
 from ._IsmrmrdRawTestData import IsmrmrdRawTestData
 from ._Dicom2DTestImage import Dicom2DTestImage
+from ._PulseqRadialTestSeq import PulseqRadialTestSeq
