+++ tests/data/test_trajectory.py
@@ -19,39 +19,8 @@ from mrpro.data.enums import TrajType
 
 from tests import RandomGenerator
 from tests.conftest import COMMON_MR_TRAJECTORIES
+from tests.conftest import create_traj
 
-
-def create_uniform_traj(nk, k_shape):
-    """Create a tensor of uniform points with predefined shape nk."""
-    kidx = torch.where(torch.tensor(nk[1:]) > 1)[0]
-    if len(kidx) > 1:
-        raise ValueError('nk is allowed to have at most one non-singleton dimension')
-    if len(kidx) >= 1:
-        # kidx+1 because we searched in nk[1:]
-        n_kpoints = nk[kidx + 1]
-        # kidx+2 because k_shape also includes coils dimensions
-        k = torch.linspace(-k_shape[kidx + 2] // 2, k_shape[kidx + 2] // 2 - 1, n_kpoints, dtype=torch.float32)
-        views = [1 if i != n_kpoints else -1 for i in nk]
-        k = k.view(*views).expand(list(nk))
-    else:
-        k = torch.zeros(nk)
-    return k
-
-
-def create_traj(k_shape, nkx, nky, nkz, sx, sy, sz):
-    """Create trajectory with random entries."""
-    random_generator = RandomGenerator(seed=0)
-    k_list = []
-    for spacing, nk in zip([sz, sy, sx], [nkz, nky, nkx], strict=True):
-        if spacing == 'nuf':
-            k = random_generator.float32_tensor(size=nk)
-        elif spacing == 'uf':
-            k = create_uniform_traj(nk, k_shape=k_shape)
-        elif spacing == 'z':
-            k = torch.zeros(nk)
-        k_list.append(k)
-    trajectory = KTrajectory(k_list[0], k_list[1], k_list[2], repeat_detection_tolerance=None)
-    return trajectory
 
 
 def test_trajectory_repeat_detection_tol(cartesian_grid):
