+++ tests/operators/models/test_molli.py
@@ -17,7 +17,7 @@
 import pytest
 import torch
 from mrpro.operators.models import MOLLI
-from tests.conftest import SHAPE_VARIATIONS_SIGNAL_MODELS, create_parameter_tensor_tuples
+from tests.operators.models.conftest import SHAPE_VARIATIONS_SIGNAL_MODELS, create_parameter_tensor_tuples
 
 
 @pytest.mark.parametrize(
