+++ tests/operators/models/test_mono_exponential_decay.py
@@ -17,7 +17,7 @@
 import pytest
 import torch
 from mrpro.operators.models import MonoExponentialDecay
-from tests.conftest import SHAPE_VARIATIONS_SIGNAL_MODELS, create_parameter_tensor_tuples
+from tests.operators.models.conftest import SHAPE_VARIATIONS_SIGNAL_MODELS, create_parameter_tensor_tuples
 
 
 @pytest.mark.parametrize(
