+++ tests/operators/test_cartesian_sampling_op.py
@@ -18,7 +18,7 @@ from mrpro.data import KTrajectory, SpatialDimension
 from mrpro.operators import CartesianSamplingOp
 
 from tests import RandomGenerator
-from tests.data.test_trajectory import create_traj
+from tests.conftest import create_traj
 from tests.helper import dotproduct_adjointness_test
