+++ tests/operators/test_fourier_op.py
@@ -18,7 +18,7 @@ from mrpro.operators import FourierOp
 
 from tests import RandomGenerator
 from tests.conftest import COMMON_MR_TRAJECTORIES
-from tests.data.test_trajectory import create_traj
+from tests.conftest import create_traj
 from tests.helper import dotproduct_adjointness_test
